改进linux下启动脚本

+++ src/main/scripts/shutdown.sh
@@ -3,6 +3,7 @@
 P_DIR=$(cd `pwd`/..; pwd)
 
 TEMP_DIR=$P_DIR/temp
-
-cat $TEMP_DIR/pid | xargs kill
-rm -rf $TEMP_DIR/pid
+if [ -e "$TEMP_DIR/pid" ]; then
+  cat $TEMP_DIR/pid | xargs kill
+  rm -rf $TEMP_DIR/pid
+fi
